feat(Client): Require and provide default for response_types
The OpenID Connect specification makes it clear that this property is optional, and does not need to be specified in order to register a client. However, it provides a desired default value and behaviour.

Marking this property as required in the schema and providing a default value effectively makes it optional, as the default value is used whenever a desired value is not explicitly specified. This falls in line with the defined behaviour in the specification.