feat(migrations): draft of formalized migrations
Upgrading to new versions of code often requires changes to an existing
database. Currently, we've made the code to automate migration in the
`boot/database.js` file. The size of this file is growing and will
become unmaintainable as we add migration code for future versions. This
commit expresses a proposed way of formalizing migrations. With this
strategy, we'll be able to add a new module that handles migration for
each new version. In addition, separating the migration code in this way
makes it testable in isolation.