Ensure AD/LDAP roles exist before assigning to user
Prior, a user with AD/LDAP groups without equivalent roles in Anvil
Connect would still have those roles assigned, creating user-to-role
relations in the database that referred to non-existent roles. This in
turn caused errors elsewhere in the server whenever the roles were
enumerated, in particular when attempting to dereference the `name`
property of a null role.