Add coveralls badge and improve build script.
The introduction of the coveralls build somewhat unintentionally
had all tests running at once. This used to fail which was the reason
that the coverage.sh script was introduced so that the coverage would
run separately for unit and integration tests and the results were combined.

However it should really be possible to run all tests at once, so that
it makes sense to do so in the build. This should allow to see regressions
in this respect.

The coverage script itself is kept as it can still be useful during
development if only to see how one can run the commands in a shell without
installing istanbul or mocha globally.

Finally the travis script now sends coverage data in a second step to
coveralls with the intent to easier distinguish cases where the tests fail and
where coveralls fails.